+++ lib/BuzzerManager/BuzzerManager.h
@@ -5,10 +5,10 @@
 #ifndef BUZZERMANAGER_H
 #define BUZZERMANAGER_H
 
-#include "Arduino.h"
+#include <Arduino.h>
 
 #include "Settings.h"
-#include "PWMSine.h"
+#include <PWMSine.h>
 
 
 /*!
@@ -69,6 +69,8 @@ public:
         @param sine An initialized PWMSine object. This is only used if
                     #active_buzzer is not defined.
     */
+    // cppcheck-suppress uninitMemberVar symbolName=BuzzerManager::period_
+    // cppcheck-suppress uninitMemberVar symbolName=BuzzerManager::freq_
     BuzzerManager(byte pin, PWMSine& sine) : pin_(pin), sine_(sine) {};
 
     void loop();