+++ test/test_embedded/test_alarm.cpp
@@ -1,45 +1,14 @@
-#line 2 "tests.ino"
+#include <Arduino.h>
+#include <unity.h>
 
-// Unit tests
-// https://github.com/bxparks/AUnit
-
-
-#include <AUnit.h>
-
-// Symlink src to ../src
-#include "src/AlarmClock/Settings.h"
-#include "src/AlarmClock/Constants.h"
-#include "src/AlarmClock/Alarm.h"
-#include "src/AlarmClock/DaysOfWeek.h"
-#include "src/AlarmClock/CountdownTimer.h"
-
-#include "Alarm_mockups.h"
 #include <RTClib.h>
+#include "Alarm_mockups.h"
+#include <Alarm.h>
+#include <DaysOfWeek.h>
+#include <CountdownTimer.h>
 
 
-void setup()
-{
-    delay(500); // wait for stability on some boards to prevent garbage Serial
-    Serial.begin(115200);
-    while(!Serial);
-    Serial.println();
-    Serial.println();
-    Serial.println("AlarmClock tests");
-    Serial.print("Compiled on ");
-    Serial.print(__DATE__);
-    Serial.print(" at ");
-    Serial.println(__TIME__);
-    Serial.println();
-}
-
-
-void loop()
-{
-    aunit::TestRunner::run();
-}
-
-
-test(Alarm_trigger)
+void test_Alarm_trigger()
 {
     MockupPWMDimmer ambientDimmer;
 
@@ -57,14 +26,14 @@ test(Alarm_trigger)
         TestAlarm alarm;  // needs to be here to reset prev_activation_millis
         alarm.SetHardware(&lamp, &ambientDimmer, &buzzer,
                           writeEEPROM, activation_callback, stop_callback);
-        assertTrue(alarm.set_time(12, 13));
-        assertTrue(alarm.set_enabled(Single));
+        TEST_ASSERT_TRUE(alarm.set_time(12, 13));
+        TEST_ASSERT_TRUE(alarm.set_enabled(Single));
         DaysOfWeek dow;
-        dow.DaysOfWeek = 0xFE;
-        assertTrue(alarm.set_days_of_week(dow));
-        assertTrue(alarm.set_snooze(1, 2));
-        assertTrue(alarm.set_signalization(80, true, true));
-        assertTrue(alarm.set_inhibit(false));
+        dow.days_of_week = 0xFE;
+        TEST_ASSERT_TRUE(alarm.set_days_of_week(dow));
+        TEST_ASSERT_TRUE(alarm.set_snooze(1, 2));
+        TEST_ASSERT_TRUE(alarm.set_signalization(80, true, true));
+        TEST_ASSERT_TRUE(alarm.set_inhibit(false));
 
         reset_alarm_mockups();
 
@@ -80,19 +49,19 @@ test(Alarm_trigger)
                 {
                     alarm.loop(time);
 
-                    assertFalse(activated);
-                    assertFalse(lamp_status);
+                    TEST_ASSERT_FALSE(activated);
+                    TEST_ASSERT_FALSE(lamp_status);
 
                     if (time.hour() == 12 && time.minute() == 13 && time.second() < 30)
                     {
-                        assertEqual(alarm.get_enabled(), Off);  // single
+                        TEST_ASSERT_EQUAL_INT(Off, alarm.get_enabled());  // single
                         activation_time = true;
                     }
 
                     reset_alarm_mockups();
                     time = time + TimeSpan(30);
                 }
-                assertTrue(activation_time);
+                TEST_ASSERT_TRUE(activation_time);
                 }
                 break;
 
@@ -105,17 +74,17 @@ test(Alarm_trigger)
 
                     if (time.hour() == 12 && time.minute() == 13 && time.second() < 30)
                     {
-                        assertTrue(activated);
-                        assertTrue(lamp_status);
-                        assertTrue(buzzer.get_status());
-                        assertEqual(alarm.get_enabled(), Off);  // single
+                        TEST_ASSERT_TRUE(activated);
+                        TEST_ASSERT_TRUE(lamp_status);
+                        TEST_ASSERT_TRUE(buzzer.get_status());
+                        TEST_ASSERT_EQUAL_INT(Off, alarm.get_enabled());  // single
 
                         activation_time = true;
                     }
                     else
                     {
-                        assertFalse(activated);
-                        assertFalse(lamp_status);
+                        TEST_ASSERT_FALSE(activated);
+                        TEST_ASSERT_FALSE(lamp_status);
                     }
 
                     reset_alarm_mockups();
@@ -123,13 +92,13 @@ test(Alarm_trigger)
                     time = time + TimeSpan(30);
                 }
                 // In case the test code was wrong
-                assertTrue(activation_time);
+                TEST_ASSERT_TRUE(activation_time);
                 }
                 break;
 
             case skip:
                 {
-                assertTrue(alarm.set_enabled(Skip));
+                TEST_ASSERT_TRUE(alarm.set_enabled(Skip));
 
                 bool activation_time = false;
                 while (time < DateTime(2020, 1, 1, 13, 15, 0))
@@ -138,17 +107,17 @@ test(Alarm_trigger)
 
                     if (time.hour() == 12 && time.minute() == 13 && time.second() < 30)
                     {
-                        assertFalse(activated);
-                        assertFalse(lamp_status);
-                        assertFalse(buzzer.get_status());
-                        assertEqual(alarm.get_enabled(), Repeat);  // Skip
+                        TEST_ASSERT_FALSE(activated);
+                        TEST_ASSERT_FALSE(lamp_status);
+                        TEST_ASSERT_FALSE(buzzer.get_status());
+                        TEST_ASSERT_EQUAL_INT(Repeat, alarm.get_enabled());  // Skip
 
                         activation_time = true;
                     }
                     else
                     {
-                        assertFalse(activated);
-                        assertFalse(lamp_status);
+                        TEST_ASSERT_FALSE(activated);
+                        TEST_ASSERT_FALSE(lamp_status);
                     }
 
                     reset_alarm_mockups();
@@ -156,18 +125,17 @@ test(Alarm_trigger)
                     time = time + TimeSpan(30);
                 }
                 // In case the test code was wrong
-                assertTrue(activation_time);
+                TEST_ASSERT_TRUE(activation_time);
                 }
                 break;
         }
 
         myTest = (test)(myTest + 1);
     }
-
 }
 
 
-test(Alarm_snooze)
+void test_Alarm_snooze()
 {
     TestAlarm alarm;
 
@@ -175,13 +143,13 @@ test(Alarm_snooze)
     alarm.SetHardware(&lamp, &ambientDimmer, &buzzer,
                       writeEEPROM, activation_callback, stop_callback);
 
-    assertTrue(alarm.set_time(12, 13));
-    assertTrue(alarm.set_enabled(Single));
+    TEST_ASSERT_TRUE(alarm.set_time(12, 13));
+    TEST_ASSERT_TRUE(alarm.set_enabled(Single));
     DaysOfWeek dow;
-    dow.DaysOfWeek = 0xFE;
-    assertTrue(alarm.set_days_of_week(dow));
-    assertTrue(alarm.set_snooze(0, 2));
-    assertTrue(alarm.set_signalization(80, true, true));
+    dow.days_of_week = 0xFE;
+    TEST_ASSERT_TRUE(alarm.set_days_of_week(dow));
+    TEST_ASSERT_TRUE(alarm.set_snooze(0, 2));
+    TEST_ASSERT_TRUE(alarm.set_signalization(80, true, true));
 
     reset_alarm_mockups();
 
@@ -192,18 +160,24 @@ test(Alarm_snooze)
         //activated = false;
         alarm.loop(time);
 
-        assertTrue(activated);
-        assertTrue(lamp_status);
-        assertTrue(buzzer.get_status());
+        TEST_ASSERT_TRUE(activated);
+        TEST_ASSERT_TRUE(lamp_status);
+        TEST_ASSERT_TRUE(buzzer.get_status());
 
         alarm.ButtonSnooze();
-        if (snooze_remaining != 0) assertFalse(lamp_status);
-        else assertTrue(lamp_status);
+        if (snooze_remaining != 0)
+        {
+            TEST_ASSERT_FALSE(lamp_status);
+        }
+        else
+        {
+            TEST_ASSERT_TRUE(lamp_status);
+        }
     }
 }
 
 
-test(Alarm_ambient)
+void test_Alarm_ambient()
 {
     enum ambientTest {
         inhibit = 0,
@@ -221,14 +195,14 @@ test(Alarm_ambient)
         TestAlarm alarm;  // needs to be here to reset prev_activation_millis
         alarm.SetHardware(&lamp, &ambientDimmer, &buzzer,
                           writeEEPROM, activation_callback, stop_callback);
-        assertTrue(alarm.set_time(12, 13));
-        assertTrue(alarm.set_enabled(Repeat));
+        TEST_ASSERT_TRUE(alarm.set_time(12, 13));
+        TEST_ASSERT_TRUE(alarm.set_enabled(Repeat));
         DaysOfWeek dow;
-        dow.DaysOfWeek = 0xFE;
-        assertTrue(alarm.set_days_of_week(dow));
-        assertTrue(alarm.set_snooze(1, 2));
-        assertTrue(alarm.set_signalization(80, true, true));
-        assertTrue(alarm.set_inhibit(false));
+        dow.days_of_week = 0xFE;
+        TEST_ASSERT_TRUE(alarm.set_days_of_week(dow));
+        TEST_ASSERT_TRUE(alarm.set_snooze(1, 2));
+        TEST_ASSERT_TRUE(alarm.set_signalization(80, true, true));
+        TEST_ASSERT_TRUE(alarm.set_inhibit(false));
 
         reset_alarm_mockups();
 
@@ -245,10 +219,10 @@ test(Alarm_ambient)
                 {
                     alarm.loop(time);
 
-                    assertFalse(activated);
-                    assertFalse(lamp_status);
-                    assertFalse(alarm.test_get_ambient_status());
-                    assertEqual(ambientDimmer.get_stop(), 0);
+                    TEST_ASSERT_FALSE(activated);
+                    TEST_ASSERT_FALSE(lamp_status);
+                    TEST_ASSERT_FALSE(alarm.test_get_ambient_status());
+                    TEST_ASSERT_EQUAL_INT(0, ambientDimmer.get_stop());
 
                     if (time.hour() == 12 && time.minute() == 13 && time.second() < 30)
                     {
@@ -265,8 +239,8 @@ test(Alarm_ambient)
                     time = time + TimeSpan(30);
                 }
                 // In case the test code was wrong
-                assertTrue(activation_time);
-                assertTrue(ambient_activation_time);
+                TEST_ASSERT_TRUE(activation_time);
+                TEST_ASSERT_TRUE(ambient_activation_time);
                 }
                 break;
 
@@ -282,28 +256,28 @@ test(Alarm_ambient)
                     DateTime temptime = time + TimeSpan(long(Alarm_ambient_dimming_duration / 1000UL));
                     if (temptime.hour() == 12 && temptime.minute() == 13 && temptime.second() < 30)
                     {
-                        assertEqual(ambientDimmer.get_stop(), 80);
-                        assertTrue(alarm.test_get_ambient_status());
+                        TEST_ASSERT_EQUAL_INT(80, ambientDimmer.get_stop());
+                        TEST_ASSERT_TRUE(alarm.test_get_ambient_status());
                         ambient_activation_time = true;
                     }
                     else
                     {
-                        assertEqual(ambientDimmer.get_stop(), 0);
-                        assertFalse(alarm.test_get_ambient_status());
+                        TEST_ASSERT_EQUAL_INT(0, ambientDimmer.get_stop());
+                        TEST_ASSERT_FALSE(alarm.test_get_ambient_status());
                     }
 
                     if (time.hour() == 12 && time.minute() == 13 && time.second() < 30)
                     {
-                        assertTrue(activated);
-                        assertTrue(lamp_status);
-                        assertTrue(buzzer.get_status());
+                        TEST_ASSERT_TRUE(activated);
+                        TEST_ASSERT_TRUE(lamp_status);
+                        TEST_ASSERT_TRUE(buzzer.get_status());
 
                         activation_time = true;
                     }
                     else
                     {
-                        assertFalse(activated);
-                        assertFalse(lamp_status);
+                        TEST_ASSERT_FALSE(activated);
+                        TEST_ASSERT_FALSE(lamp_status);
                     }
 
                     reset_alarm_mockups();
@@ -311,8 +285,8 @@ test(Alarm_ambient)
                     time = time + TimeSpan(30);
                 }
                 // In case the test code was wrong
-                assertTrue(activation_time);
-                assertTrue(ambient_activation_time);
+                TEST_ASSERT_TRUE(activation_time);
+                TEST_ASSERT_TRUE(ambient_activation_time);
                 }
                 break;
 
@@ -320,15 +294,15 @@ test(Alarm_ambient)
                 {
                 bool activation_time = false;
                 bool ambient_activation_time = false;
-                assertTrue(alarm.set_enabled(Skip));
+                TEST_ASSERT_TRUE(alarm.set_enabled(Skip));
                 while (time < DateTime(2020, 1, 1, 13, 15, 0))
                 {
                     alarm.loop(time);
 
-                    assertFalse(activated);
-                    assertFalse(lamp_status);
-                    assertFalse(alarm.test_get_ambient_status());
-                    assertEqual(ambientDimmer.get_stop(), 0);
+                    TEST_ASSERT_FALSE(activated);
+                    TEST_ASSERT_FALSE(lamp_status);
+                    TEST_ASSERT_FALSE(alarm.test_get_ambient_status());
+                    TEST_ASSERT_EQUAL_INT(0, ambientDimmer.get_stop());
 
                     if (time.hour() == 12 && time.minute() == 13 && time.second() < 30)
                     {
@@ -345,8 +319,8 @@ test(Alarm_ambient)
                     time = time + TimeSpan(30);
                 }
                 // In case the test code was wrong
-                assertTrue(activation_time);
-                assertTrue(ambient_activation_time);
+                TEST_ASSERT_TRUE(activation_time);
+                TEST_ASSERT_TRUE(ambient_activation_time);
                 }
                 break;
 
@@ -356,7 +330,7 @@ test(Alarm_ambient)
 }
 
 
-test(Alarm_EEPROM_read)
+void test_Alarm_EEPROM_read()
 {
     TestAlarm alarm;
     MockupPWMDimmer ambientDimmer;
@@ -367,70 +341,70 @@ test(Alarm_EEPROM_read)
     byte data1[Alarm::EEPROM_length] = {
         EEPROM_alarms_id + 1, 23, 59, 3, 0x08, 99, 9, 80, 1, 0
     };
-    assertFalse(alarm.ReadEEPROM(data1));
+    TEST_ASSERT_FALSE(alarm.ReadEEPROM(data1));
 
 
     // Bad hours
     byte data2[Alarm::EEPROM_length] = {
         EEPROM_alarms_id, 24, 59, 3, 0x08, 99, 9, 80, 1, 0
     };
-    assertFalse(alarm.ReadEEPROM(data2));
+    TEST_ASSERT_FALSE(alarm.ReadEEPROM(data2));
 
     // Bad minutes
     byte data3[Alarm::EEPROM_length] = {
         EEPROM_alarms_id, 23, 60, 3, 0x08, 99, 9, 80, 1, 0
     };
-    assertFalse(alarm.ReadEEPROM(data3));
+    TEST_ASSERT_FALSE(alarm.ReadEEPROM(data3));
 
     // Bad enabled
     byte data4[Alarm::EEPROM_length] = {
         EEPROM_alarms_id, 23, 59, 4, 0x08, 99, 9, 80, 1, 0
     };
-    assertFalse(alarm.ReadEEPROM(data4));
+    TEST_ASSERT_FALSE(alarm.ReadEEPROM(data4));
 
     // Bad snooze time
     byte data5[Alarm::EEPROM_length] = {
         EEPROM_alarms_id, 23, 59, 3, 0x08, 100, 9, 80, 1, 0
     };
-    assertFalse(alarm.ReadEEPROM(data5));
+    TEST_ASSERT_FALSE(alarm.ReadEEPROM(data5));
 
     // Bad snooze count
     byte data6[Alarm::EEPROM_length] = {
         EEPROM_alarms_id, 23, 59, 3, 0x08, 99, 10, 80, 1, 0
     };
-    assertFalse(alarm.ReadEEPROM(data6));
+    TEST_ASSERT_FALSE(alarm.ReadEEPROM(data6));
 
     // Valid alarm
     byte data7[Alarm::EEPROM_length] = {
         EEPROM_alarms_id, 23, 59, 3, 0x08, 99, 9, 80, 1, 0
     };
-    assertTrue(alarm.ReadEEPROM(data7));
-    assertEqual(alarm.get_time().hours, 23);
-    assertEqual(alarm.get_time().minutes, 59);
-    assertEqual(alarm.get_enabled(), Skip);
-    assertEqual(alarm.get_snooze().time_minutes, 99);
-    assertEqual(alarm.get_snooze().count, 9);
-    assertEqual(alarm.get_signalization().ambient, 80);
-    assertEqual(alarm.get_signalization().lamp, true);
-    assertFalse(alarm.get_signalization().buzzer);
+    TEST_ASSERT_TRUE(alarm.ReadEEPROM(data7));
+    TEST_ASSERT_EQUAL_INT(23, alarm.get_time().hours);
+    TEST_ASSERT_EQUAL_INT(59, alarm.get_time().minutes);
+    TEST_ASSERT_EQUAL_INT(Skip, alarm.get_enabled());
+    TEST_ASSERT_EQUAL_INT(99, alarm.get_snooze().time_minutes);
+    TEST_ASSERT_EQUAL_INT(9, alarm.get_snooze().count);
+    TEST_ASSERT_EQUAL_INT(80, alarm.get_signalization().ambient);
+    TEST_ASSERT_TRUE(alarm.get_signalization().lamp);
+    TEST_ASSERT_FALSE(alarm.get_signalization().buzzer);
 }
 
 
-test(Alarm_EEPROM_write)
+void test_Alarm_EEPROM_write()
 {
     TestAlarm alarm;
     MockupPWMDimmer ambientDimmer;
     alarm.SetHardware(&lamp, &ambientDimmer, &buzzer,
                       writeEEPROM, activation_callback, stop_callback);
 
-    assertTrue(alarm.set_time(23, 59));
-    assertTrue(alarm.set_enabled(Repeat));
+    TEST_ASSERT_TRUE(alarm.set_time(23, 59));
+    TEST_ASSERT_TRUE(alarm.set_enabled(Repeat));
     DaysOfWeek dow;
-    dow.DaysOfWeek = 0x08;
-    assertTrue(alarm.set_days_of_week(dow));
-    assertTrue(alarm.set_snooze(99, 9));
-    assertTrue(alarm.set_signalization(80, true, false));
-    assertTrue(alarm.set_inhibit(false));
+    dow.days_of_week = 0x08;
+    TEST_ASSERT_TRUE(alarm.set_days_of_week(dow));
+    TEST_ASSERT_TRUE(alarm.set_snooze(99, 9));
+    TEST_ASSERT_TRUE(alarm.set_signalization(80, true, false));
+    TEST_ASSERT_TRUE(alarm.set_inhibit(false));
 
     byte * data = alarm.WriteEPROM();
 
@@ -440,12 +414,12 @@ test(Alarm_EEPROM_write)
 
     for (byte i = 0; i < Alarm::EEPROM_length; i++)
     {
-        assertEqual(data[i], correct_data[i]);
+        TEST_ASSERT_EQUAL_INT(correct_data[i], data[i]);
     }
 }
 
 
-test(Timer)
+void test_Timer()
 {
     MockupPWMDimmer ambientDimmer;
     CountdownTimer timer(ambientDimmer, lamp, buzzer);
@@ -465,31 +439,64 @@ test(Timer)
         // 11:31:32 - timer.start() causes time_left to immediately decrement.
         if (time.hour() == 11 && time.minute() == 31 && time.second() == 32)
         {
-            assertTrue(lamp_status);
-            assertTrue(buzzer.get_status());
-            assertEqual(ambientDimmer.get_stop(), 120);
+            TEST_ASSERT_TRUE(lamp_status);
+            TEST_ASSERT_TRUE(buzzer.get_status());
+            TEST_ASSERT_EQUAL_INT(120, ambientDimmer.get_stop());
 
             timer.ButtonStop();
 
-            assertEqual(ambientDimmer.get_stop(), 0);
-            assertFalse(lamp_status);
-            assertFalse(buzzer.get_status());
+            TEST_ASSERT_EQUAL_INT(0, ambientDimmer.get_stop());
+            TEST_ASSERT_FALSE(lamp_status);
+            TEST_ASSERT_FALSE(buzzer.get_status());
 
             activation_time = true;
         }
         else
         {
-            if (!activation_time) assertTrue(timer.get_running());
-            else assertFalse(timer.get_running());
-            assertFalse(lamp_status);
-            assertFalse(buzzer.get_status());
-            assertEqual(ambientDimmer.get_stop(), 0);
+            if (!activation_time)
+            {
+                TEST_ASSERT_TRUE(timer.get_running());
+            }
+            else
+            {
+                TEST_ASSERT_FALSE(timer.get_running());
+            }
+            TEST_ASSERT_FALSE(lamp_status);
+            TEST_ASSERT_FALSE(buzzer.get_status());
+            TEST_ASSERT_EQUAL_INT(0, ambientDimmer.get_stop());
         }
 
         reset_alarm_mockups();
         time = time + TimeSpan(1);
     }
     // In case the test code was wrong
-    assertTrue(activation_time);
+    TEST_ASSERT_TRUE(activation_time);
 
 }
+
+
+void setup()
+{
+    // NOTE!!! Wait for >2 secs
+    // if board doesn't support software reset via Serial.DTR/RTS
+    delay(2000);
+
+    UNITY_BEGIN();
+    RUN_TEST(test_Alarm_trigger);
+    RUN_TEST(test_Alarm_snooze);
+    RUN_TEST(test_Alarm_ambient);
+    RUN_TEST(test_Alarm_EEPROM_read);
+    RUN_TEST(test_Alarm_EEPROM_write);
+    RUN_TEST(test_Timer);
+    UNITY_END();
+}
+
+
+void loop()
+{
+    digitalWrite(LED_BUILTIN, HIGH);
+    delay(100);
+    digitalWrite(LED_BUILTIN, LOW);
+    delay(500);
+}
+