+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for melodies.
 - CLI now sends the BEL character (`0x07`) every time the state of the
   hardware (ambient, lamp) or settings (inhibit) changes.
+- Native unit tests.
 
 ### Changed
 - EEPROM data storage format. The section that contains alarms has been moved.
@@ -33,6 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     ```
 - Default value of `Alarm_ambient_fade_out_duration` is now 10 seconds instead
   of 2.
+- Firmware is now compiled using PlatformIO instead of Arduino IDE.
+  A `Makefile` has been added.
 
 ### Removed
 - DEBUG EEPROM dump after boot - this is no longer useful now that we have
@@ -41,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Potential out-of-bounds write in SerialCLI.
+- Uninitialized variable in PWMDimmer.
 
 
 ## [0.5.0] - 2021-11-10